UPSTREAM: fs/fat: Reduce stack usage
We have limited stack in SPL builds.  Drop itrblock and move to
malloc/free of itr to move this off of the stack.  As part of this fix a
double-free issue in fat_size().

Change-Id: I2d8baa3a0b6ffa0f9591366a095d5e315c41fde8
Signed-off-by: Tom Rini <trini@konsulko.com>
---
Rework to use malloc/free as moving this to a global overflows some SH
targets.
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>
(cherry picked from commit 2460098cffacd18729262e3ed36656e6943783ed)