usb: gadget: dwc2: avoid reset core before devices connected
Theoretically, the UDC register need not reconfig when GINTSTS.USBRst
is set, and only do core reset if the devices was connected before.
This change adds _connected_ flag to check the devices was really
connected or not. As an optimization, reset device address to zero
while GINTSTS.USBRst is set.

This patch fix usb connect failed when continuously perform the
'fastboot reboot-bootloader' command.

Change-Id: I6a78228e147d2274329d922ac3f3ffef19492e7e
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>